[pipeline] Move isolating-pool executor helpers to _common/_convert
Pure refactor, no behavior change. Moves the executor-classification helpers `_is_isolating_pool` and `_is_interpreter_pool` — and the `_INTERPRETER_POOL_CLASS` version guard they depend on — from `spdl/pipeline/_fuse.py` into `spdl/pipeline/_common/_convert.py`, next to the existing `_is_process_pool` they build on.

`_common/_convert` sits low in the import graph, so hosting the helpers there lets other modules (notably `spdl/pipeline/defs`) reuse them without importing `_fuse` and creating an import cycle. `_fuse` now imports `_is_isolating_pool` from `_common/_convert` instead of defining it.

Extracted from https://github.com/facebookresearch/spdl/pull/1579
