+++ src/spdl/pipeline/_common/_convert.py
@@ -9,6 +9,8 @@
 __all__ = [
     "convert_to_async",
     "_is_process_pool",
+    "_is_interpreter_pool",
+    "_is_isolating_pool",
     "_to_async_gen",
 ]
 
@@ -38,6 +40,63 @@ def _is_process_pool(executor: Executor | type[Executor] | None) -> bool:
     if owner is not None:
         return _is_process_pool(owner)
     return False
+
+
+# Declared before the branch so its type stays ``type[Executor] | None`` on every Python
+# version; otherwise, on versions without ``InterpreterPoolExecutor`` the ``else`` assignment
+# narrows it to ``None`` and the isinstance/issubclass checks below fail to type-check.
+_INTERPRETER_POOL_CLASS: type[Executor] | None
+if sys.version_info >= (3, 14):
+    from concurrent.futures.interpreter import (  # pyre-ignore[21]
+        InterpreterPoolExecutor,
+    )
+
+    _INTERPRETER_POOL_CLASS = InterpreterPoolExecutor
+else:
+    _INTERPRETER_POOL_CLASS = None
+
+
+def _is_interpreter_pool(executor: Executor | type[Executor] | None) -> bool:
+    """Check whether ``executor`` is or wraps an ``InterpreterPoolExecutor``.
+
+    Mirrors :py:func:`_is_process_pool`: matches the stdlib ``InterpreterPoolExecutor``
+    (Python 3.14+) directly, SPDL pool wrappers via their ``_pool_executor_class``, and a
+    ``PriorityExecutorEntrypoint`` via its ``_owner``.
+
+    Args:
+        executor: The executor (or executor class, or ``None``) to inspect.
+
+    Returns:
+        ``True`` if the executor isolates work in a subinterpreter, else ``False``. Always
+        ``False`` on Python versions without ``InterpreterPoolExecutor``.
+    """
+    if _INTERPRETER_POOL_CLASS is None or executor is None:
+        return False
+    if isinstance(executor, _INTERPRETER_POOL_CLASS):
+        return True
+    pool_cls = getattr(executor, "_pool_executor_class", None)
+    if pool_cls is not None:
+        return issubclass(pool_cls, _INTERPRETER_POOL_CLASS)
+    owner = getattr(executor, "_owner", None)
+    if owner is not None:
+        return _is_interpreter_pool(owner)
+    return False
+
+
+def _is_isolating_pool(executor: Executor | type[Executor] | None) -> bool:
+    """Check whether ``executor`` runs work in a separate process or subinterpreter.
+
+    These are exactly the executors whose inter-stage handoff crosses an IPC / pickling
+    boundary, and therefore the executors that fusion targets. Thread pools (which share
+    address space) and ``None`` (the default thread pool) are not isolating.
+
+    Args:
+        executor: The executor (or executor class, or ``None``) to inspect.
+
+    Returns:
+        ``True`` if the executor is a process pool or interpreter pool, else ``False``.
+    """
+    return _is_process_pool(executor) or _is_interpreter_pool(executor)
 
 
 def _func_in_subprocess(func: Callable[[T], U], item: T) -> U:
